requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
selenium==3.141.0
python-dotenv~=0.15.0